+++ mysql/CHANGELOG.md
@@ -2,7 +2,18 @@
 
 <!-- towncrier release notes start -->
 
-## 14.4.0 / 2024-12-26
+## 14.5.0 / 2025-01-25
+
+***Added***:
+
+* Added an optional optimization to MySQL statement metrics collection to only query for queries that have run since the last check collection. ([#19321](https://github.com/DataDog/integrations-core/pull/19321))
+* Emit index usage and index metrics from mysql integration ([#19383](https://github.com/DataDog/integrations-core/pull/19383))
+
+***Fixed***:
+
+* Bump datadog-checks-base version ([#19478](https://github.com/DataDog/integrations-core/pull/19478))
+
+## 14.4.0 / 2024-12-26 / Agent 7.62.0
 
 ***Added***: